rename native-build-settings -> native-build-step
