+++ pants.ini
@@ -406,17 +406,17 @@ fragment: True
 [sitegen]
 config_path: src/docs/docsite.json
 
+# TODO(#6848, #7122): A gcc toolchain is currently required to build the tensorflow_custom_op example. This
+# should be removed when we can fully support both toolchains, and select between them with constraints.
+[native-build-step]
+toolchain_variant: gnu
+
 [native-build-step.cpp-compile-settings]
 compiler_option_sets_enabled_args: {
     # TensorFlow custom operators cannot be built against the C++11 ABI yet: see
     # https://www.tensorflow.org/guide/extend/op.
     'glibcxx_use_old_abi': ['-D_GLIBCXX_USE_CXX11_ABI=0'],
   }
-
-# TODO(#6848, #7122): A gcc toolchain is currently required to build the tensorflow_custom_op example. This
-# should be removed when we can fully support both toolchains, and select between them with constraints.
-[native-build-step]
-toolchain_variant: gnu
 
 [libc]
 enable_libc_search: True
